changed quantity field in doughnuts table from float to int

+++ db/migrate/20211101054516_add_details_to_doughnuts.rb
@@ -1,5 +1,5 @@
 class AddDetailsToDoughnuts < ActiveRecord::Migration[6.1]
   def change
-    add_column :doughnuts, :quantity, :float
+    add_column :doughnuts, :quantity, :int
   end
 end
